requirements.txt
asgiref==3.6.0
Django==4.1.7
django-restframework==0.0.1
sqlparse==0.4.3
tzdata==2022.7
